fast-api/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
sqlalchemy==2.0.38
pymysql==1.1.1
openai
